Add research whitepaper on NLP classifiers for trading
This document presents a comparative study of various NLP classifiers for event-driven trading in finance, emphasizing the need for heterogeneous approaches to accurately classify news headlines. This is part of my applied research for a unique quant finance / fintech course that I have been developing the past few months for educational and industry level research.